rename shared components folder
