+++ client/src/pages/shared_components/FeatureTable.jsx
